% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_vig.R
\name{analyze_repo_vignette}
\alias{analyze_repo_vignette}
\title{obtain information to support content analysis of a package via its vignette}
\usage{
analyze_repo_vignette(repourl, keyname = "ANTHROPIC_API_KEY")
}
\arguments{
\item{repourl}{character URL to github repository}

\item{keyname}{character environment variable name for LLM-provider API key}
}
\description{
obtain information to support content analysis of a package via its vignette
}
