% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_funcs.R
\name{classify_batch}
\alias{classify_batch}
\title{internal classification support}
\usage{
classify_batch(pkg_batch, categories, chat)
}
\description{
internal classification support
}
\keyword{internal}
