% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_funcs.R
\name{classify_packages}
\alias{classify_packages}
\title{main tool for classifying packages}
\usage{
classify_packages(
  json_file,
  n_categories = 10,
  chatfun = ellmer::chat_anthropic,
  model = "claude-sonnet-4-20250514",
  output_prefix = "packages_classified"
)
}
\arguments{
\item{json_file}{path to json document produced from list of analyze_vig outputs}

\item{n_categories}{numeric}

\item{chatfun}{defaults to ellmer::chat_anthropic}

\item{model}{defaults to "claude-sonnet-4-20250514"}

\item{output_prefix}{character prefix for emitted md, html, json}
}
\description{
main tool for classifying packages
}
