% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_support.R
\name{extract_repo_from_body}
\alias{extract_repo_from_body}
\title{given the 'body' in an issue response from Bioc contributions,
extract the repository URL}
\usage{
extract_repo_from_body(text)
}
\arguments{
\item{text}{a character string expected to include a single GitHub repo URL}
}
\description{
given the 'body' in an issue response from Bioc contributions,
extract the repository URL
}
