% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_funcs.R
\name{generate_categories}
\alias{generate_categories}
\title{produce a vector of categories based on incoming content}
\usage{
generate_categories(
  packages,
  n_categories = 10,
  chatfun = ellmer::chat_anthropic,
  model = "claude-sonnet-4-20250514"
)
}
\arguments{
\item{packages}{output of `load_packages`}

\item{n_categories}{numeric}

\item{chatfun}{defaults to ellmer::chat_anthropic}

\item{model}{defaults to "claude-sonnet-4-20250514"}
}
\value{
a list with components categories and descriptions
}
\description{
produce a vector of categories based on incoming content
}
