% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_workflow_gh.R
\name{get_github_headers}
\alias{get_github_headers}
\title{Build GitHub API headers}
\usage{
get_github_headers()
}
\value{
httr headers object
}
\description{
Build GitHub API headers
}
\keyword{internal}
