% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_funcs.R
\name{load_packages}
\alias{load_packages}
\title{import a json document built from a list of results of
analyze_repo_vignette}
\usage{
load_packages(json_file)
}
\description{
import a json document built from a list of results of
analyze_repo_vignette
}
