% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_funcs.R
\name{prepare_package_info}
\alias{prepare_package_info}
\title{internal function to manage summary}
\usage{
prepare_package_info(pkg, idx)
}
\description{
internal function to manage summary
}
\keyword{internal}
