% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig2data.R
\name{vig2data}
\alias{vig2data}
\title{Use the extract_data facility defined in ellmer's doc to obtain summary information about textual content.
Originally tailored to vignettes in bioconductor; it is newly generalized to handle any pdf, html or text in URL.}
\usage{
vig2data(
  url = "https://bioconductor.org/packages/release/bioc/html/Voyager.html",
  maxnchar = 30000,
  n_pdf_pages = 10,
  chatfunc = ellmer::chat_anthropic,
  model = "claude-sonnet-4-5-20250929",
  ...
)
}
\arguments{
\item{url}{character(1) URL for an html bioconductor vignettes}

\item{maxnchar}{numeric(1) text is truncated to a substring with this length}

\item{n_pdf_pages}{numeric(1) maximum number of pages to extract text from for pdf vignettes}

\item{chatfunc}{defaults to ellmer::chat_anthropic}

\item{model}{character(1) model for use with chat_openai, defaults to "claude-sonnet-4-5-20250929"}

\item{\dots}{passed to ellmer chat function}
}
\value{
a list with components author, topics, focused, coherence, and persuasion
}
\description{
Use the extract_data facility defined in ellmer's doc to obtain summary information about textual content.
Originally tailored to vignettes in bioconductor; it is newly generalized to handle any pdf, html or text in URL.
}
\note{
Based on code from https://cran.r-project.org/web/packages/ellmer/vignettes/structured-data.html
March 15 2025.  Requires that a suitable API key is available in environment.
}
\examples{
\donttest{
 if (interactive()) {
  # to use defaults, be sure ANTHROPIC_API_KEY is available to Sys.getenv
  tst = vig2data()
  str(tst)
 }
}
}
